mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stratum-mcp"
version = "0.1.0"
description = "MCP server exposing Stratum Sports intelligence tools"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp[cli]>=1.4",
    "httpx>=0.27",
    "uvicorn[standard]>=0.29",
    "python-dotenv>=1.0",
]

[project.scripts]
stratum-mcp = "stratum_mcp.server:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["stratum_mcp*"]
